fixed fialed and passed test
